requirements.txt
PyQt6>=6.5
PyOpenGL>=3.1
numpy>=1.24
